Refactor: make functions for icrements reusable
